Update calculator.py
修改挑战1 将int改为float

+++ 挑战1 实现个税计算机/calculator.py
@@ -31,7 +31,7 @@ if __name__ == '__main__':
         print('Parameter Error')
         exit()
     try:
-        wage = int(argv[1])
+        wage = float(argv[1])
     except ValueError:
         print('Parameter Error')
         exit()
